+++ vnext/Microsoft.ReactNative/Fabric/Composition/UiaHelpers.cpp
@@ -1,7 +1,10 @@
 #include "pch.h"
 #include "UiaHelpers.h"
 #include <Fabric/Composition/CompositionViewComponentView.h>
+#include <atlcomcli.h>
 #include <inspectable.h>
+#include "CompositionRootAutomationProvider.h"
+#include "RootComponentView.h"
 
 namespace winrt::Microsoft::ReactNative::implementation {
 
@@ -84,6 +87,57 @@ HRESULT UiaGetBoundingRectangleHelper(::Microsoft::ReactNative::ReactTaggedView 
   rect.height = clientRect.bottom - clientRect.top;
 
   return S_OK;
+}
+
+HRESULT UiaSetFocusHelper(::Microsoft::ReactNative::ReactTaggedView &view) noexcept {
+  auto strongView = view.view();
+
+  if (!strongView)
+    return UIA_E_ELEMENTNOTAVAILABLE;
+
+  auto rootCV = strongView->rootComponentView();
+  if (rootCV == nullptr)
+    return UIA_E_ELEMENTNOTAVAILABLE;
+
+  return rootCV->TrySetFocusedComponent(*strongView) ? S_OK : E_FAIL;
+}
+
+bool WasUiaPropertyAdvised(winrt::com_ptr<IRawElementProviderSimple> &providerSimple, PROPERTYID propId) noexcept {
+  auto spFragment = providerSimple.try_as<IRawElementProviderFragment>();
+  if (spFragment == nullptr)
+    return false;
+
+  winrt::com_ptr<IRawElementProviderFragmentRoot> spFragmentRoot;
+  spFragment->get_FragmentRoot(spFragmentRoot.put());
+  if (spFragmentRoot == nullptr)
+    return false;
+
+  auto rootProvider = static_cast<CompositionRootAutomationProvider *>(spFragmentRoot.get());
+
+  return rootProvider->WasPropertyAdvised(propId);
+}
+
+void UpdateUiaProperty(winrt::IInspectable provider, PROPERTYID propId, bool oldValue, bool newValue) noexcept {
+  auto spProviderSimple = provider.try_as<IRawElementProviderSimple>();
+
+  if (spProviderSimple == nullptr || oldValue == newValue || !WasUiaPropertyAdvised(spProviderSimple, propId))
+    return;
+
+  UiaRaiseAutomationPropertyChangedEvent(spProviderSimple.get(), propId, CComVariant(oldValue), CComVariant(newValue));
+}
+
+void UpdateUiaProperty(
+    winrt::IInspectable provider,
+    PROPERTYID propId,
+    const std::string &oldValue,
+    const std::string &newValue) noexcept {
+  auto spProviderSimple = provider.try_as<IRawElementProviderSimple>();
+
+  if (spProviderSimple == nullptr || oldValue == newValue || !WasUiaPropertyAdvised(spProviderSimple, propId))
+    return;
+
+  UiaRaiseAutomationPropertyChangedEvent(
+      spProviderSimple.get(), propId, CComVariant(oldValue.c_str()), CComVariant(newValue.c_str()));
 }
 
 } // namespace winrt::Microsoft::ReactNative::implementation