+++ packages/@react-native-windows/fs/CHANGELOG.md
@@ -1,8 +1,41 @@
 # Change Log - @react-native-windows/fs
 
-This log was last generated on Mon, 08 May 2023 17:41:04 GMT and should not be manually modified.
+This log was last generated on Tue, 13 Jun 2023 05:15:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0-canary.23
+
+Tue, 13 Jun 2023 05:15:41 GMT
+
+### Changes
+
+- Bump @rnw-scripts/eslint-config to v1.2.0
+
+## 0.0.0-canary.22
+
+Wed, 07 Jun 2023 05:16:11 GMT
+
+### Changes
+
+- Bump @rnw-scripts/just-task to v2.3.11
+
+## 0.0.0-canary.21
+
+Tue, 06 Jun 2023 19:06:44 GMT
+
+### Changes
+
+- integrate RN nightly 5/19/23 (tatianakapos@microsoft.com)
+- Bump @rnw-scripts/just-task to v2.3.10
+
+## 0.0.0-canary.20
+
+Tue, 16 May 2023 05:16:30 GMT
+
+### Changes
+
+- Bump @rnw-scripts/just-task to v2.3.9
 
 ## 0.0.0-canary.19