+++ packages/@react-native-windows/package-utils/CHANGELOG.md
@@ -1,8 +1,49 @@
 # Change Log - @react-native-windows/package-utils
 
-This log was last generated on Mon, 08 May 2023 17:41:04 GMT and should not be manually modified.
+This log was last generated on Tue, 13 Jun 2023 05:15:41 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0-canary.49
+
+Tue, 13 Jun 2023 05:15:41 GMT
+
+### Changes
+
+- Bump @react-native-windows/find-repo-root to v0.0.0-canary.52
+- Bump @react-native-windows/fs to v0.0.0-canary.23
+- Bump @rnw-scripts/eslint-config to v1.2.0
+
+## 0.0.0-canary.48
+
+Wed, 07 Jun 2023 05:16:11 GMT
+
+### Changes
+
+- Bump @react-native-windows/find-repo-root to v0.0.0-canary.51
+- Bump @react-native-windows/fs to v0.0.0-canary.22
+- Bump @rnw-scripts/just-task to v2.3.11
+
+## 0.0.0-canary.47
+
+Tue, 06 Jun 2023 19:06:44 GMT
+
+### Changes
+
+- integrate RN nightly 5/19/23 (tatianakapos@microsoft.com)
+- Bump @react-native-windows/find-repo-root to v0.0.0-canary.50
+- Bump @react-native-windows/fs to v0.0.0-canary.21
+- Bump @rnw-scripts/just-task to v2.3.10
+
+## 0.0.0-canary.46
+
+Tue, 16 May 2023 05:16:30 GMT
+
+### Changes
+
+- Bump @react-native-windows/find-repo-root to v0.0.0-canary.49
+- Bump @react-native-windows/fs to v0.0.0-canary.20
+- Bump @rnw-scripts/just-task to v2.3.9
 
 ## 0.0.0-canary.45